CHROMIUM: iwl7000: dbg_ini: fix bad dump size calculation
The driver initiates the size value with the size of the struct and then
adds the size of the data and checks if the size is zero so size can not
be equal to zero.

Solve this by getting the data size, check that it is not equal to zero
and only then add the struct size.

Change-Id: I9e78a96d79c6925e1b3d3d419ebd0cec228f4ba6
Signed-off-by: Shahar S Matityahu <shahar.s.matityahu@intel.com>
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
iwl7000-tree: 75ca2ac23c9cdd20b4ba242d1e4cdc9d8660bb38