Revert "CHROMIUM: scripts: Use savedefconfig to minimize kernel config"
This reverts commit eb4733ee24c6f6107ca7f4795018b035f9f86809.

The 3.x kernels have an environment variable, WIFIVERSION, that affects
which subtree is used during kconfig and build. This is a per-board
setting configured by a USE flag (eg USE=wireless34). What this means is
our kernel configs are generated and normalized based on the base wireless/
tree, but then different boards are built against wireless trees with
a different Kconfig structure. This sort of works today by luck if most of
the Kconfigs have the same names, and we're fine with the silent defaulted
values for any of the ones that are unique. My minimization change
perterbs the system more by not listing kconfigs that come by default or
via select. This allows more things to change when we swap out a completely
different tree at build time than we configured for. For instance, the
minimized version is now sensitive to changes in depends and select between
the trees, where the haystack version had a better chance to work out since
it listed everything.

Back out the splitconfig minimization for 3.x kernels, as those seem to be
held together in a very fragile way, and the benefit of consistency doesn't
outweigh the risk of destabilization.

BUG=chromium:963478
TEST=Build kernel for peppy

Change-Id: Ie745f033900e4dd00d77e2ae0d3817bc2f808580
Signed-off-by: Evan Green <evgreen@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1637430
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>
Reviewed-by: Guenter Roeck <groeck@chromium.org>